restrict singleton command dialogs #1950

+++ OneMore/Commands/Settings/SettingsCommand.cs
@@ -15,6 +15,9 @@ namespace River.OneMoreAddIn.Commands
 
 	internal class SettingsCommand : Command
 	{
+		private static bool commandIsActive = false;
+
+
 		public SettingsCommand()
 		{
 			// prevent replay
@@ -24,49 +27,59 @@ namespace River.OneMoreAddIn.Commands
 
 		public override async Task Execute(params object[] args)
 		{
-			using var dialog = new SettingsDialog(args[0] as IRibbonUI);
-			dialog.ShowDialog(owner);
+			if (commandIsActive) { return; }
+			commandIsActive = true;
 
-			if (!dialog.RestartNeeded)
+			try
 			{
-				return;
+				using var dialog = new SettingsDialog(args[0] as IRibbonUI);
+				dialog.ShowDialog(owner);
+
+				if (!dialog.RestartNeeded)
+				{
+					return;
+				}
+
+				var one = new OneNote();
+				if (one.WindowCount > 1)
+				{
+					ShowInfo(Resx.SettingsDialog_closeWindows);
+					return;
+				}
+				else if (UI.MoreMessageBox.ShowQuestion(owner,
+					Resx.SettingsDialog_Restart) != DialogResult.Yes)
+				{
+					return;
+				}
+
+				var processes = Process.GetProcessesByName("ONENOTE");
+				if (processes.Length == 0)
+				{
+					logger.WriteLine("cannot find ONENOTE process to restart");
+					return;
+				}
+
+				var path = processes[0].MainModule.FileName;
+
+				// the hidden cmd window will remain until OneNote is subsequently closed
+				// not sure how to make the cmd.exe host process close immediately...
+
+				var Info = new ProcessStartInfo
+				{
+					Arguments = $"/C taskkill /fi \"pid gt 0\" /im ONENOTE.exe && ping 127.0.0.1 -n 3 && \"{path}\"",
+					WindowStyle = ProcessWindowStyle.Hidden,
+					CreateNoWindow = true,
+					FileName = "cmd.exe"
+				};
+
+				Process.Start(Info);
+
+				await Task.Yield();
 			}
-
-			var one = new OneNote();
-			if (one.WindowCount > 1)
+			finally
 			{
-				ShowInfo(Resx.SettingsDialog_closeWindows);
-				return;
+				commandIsActive = false;
 			}
-			else if (UI.MoreMessageBox.ShowQuestion(owner,
-				Resx.SettingsDialog_Restart) != DialogResult.Yes)
-			{
-				return;
-			}
-
-			var processes = Process.GetProcessesByName("ONENOTE");
-			if (processes.Length == 0)
-			{
-				logger.WriteLine("cannot find ONENOTE process to restart");
-				return;
-			}
-
-			var path = processes[0].MainModule.FileName;
-
-			// the hidden cmd window will remain until OneNote is subsequently closed
-			// not sure how to make the cmd.exe host process close immediately...
-
-			var Info = new ProcessStartInfo
-			{
-				Arguments = $"/C taskkill /fi \"pid gt 0\" /im ONENOTE.exe && ping 127.0.0.1 -n 3 && \"{path}\"",
-				WindowStyle = ProcessWindowStyle.Hidden,
-				CreateNoWindow = true,
-				FileName = "cmd.exe"
-			};
-
-			Process.Start(Info);
-
-			await Task.Yield();
 		}
 	}
 }
